pyproject.toml
[build-system]
requires=["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "contilio"
version = "1.0.0"
authors = [
    { name = "Alex Pope", email = "alex@aroundwithalex.dev"}
]
description = "Extracts transportation data from an API"
readme = "README.md"
requires-python= ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = [
    "dateparser >=1.1.8",
    "requests>=2.28.2",
    "SQLAlchemy >= 2.0.10",
    "python-dotenv >= 1.0.0"
]
